Only locally bind `revert-buffer-function'
Setting this variable globally causes #'revert-file to stop working. I
got very confused by this, as Emacs was prompting me about a file I
was editing having changed on disk, but me being unable to reload it.

Note that this is *not* required for the following line because that
variable automatically becomes buffer-local when set.